+++ src/database/migrations/20210306225026-create-payables.js
@@ -20,7 +20,7 @@ module.exports = {
         onDelete: 'CASCADE'
       },
       price: {
-        type: Sequelize.DataTypes.INTEGER,
+        type: Sequelize.DataTypes.REAL,
         allowNull: false,
       },
       status: {
